Fix some bugs
