+++ client/client.go
@@ -23,8 +23,7 @@ import (
 )
 
 const (
-	Keystone = "keystone"
-	Noauth   = "noauth"
+	OpensdsEndpoint = "OPENSDS_ENDPOINT"
 )
 
 // Client is a struct for exposing some operations of opensds resources.
@@ -37,43 +36,6 @@ type Client struct {
 	*ReplicationMgr
 
 	cfg *Config
-}
-
-type AuthOptions interface {
-	GetTenantId() string
-}
-
-func NewKeystoneAuthOptions() *KeystoneAuthOptions {
-	return &KeystoneAuthOptions{}
-}
-
-type KeystoneAuthOptions struct {
-	IdentityEndpoint string
-	Username         string
-	UserID           string
-	Password         string
-	DomainID         string
-	DomainName       string
-	TenantID         string
-	TenantName       string
-	AllowReauth      bool
-	TokenID          string
-}
-
-func (k *KeystoneAuthOptions) GetTenantId() string {
-	return k.TenantID
-}
-
-func NewNoauthOptions(tenantId string) *NoAuthOptions {
-	return &NoAuthOptions{TenantID: tenantId}
-}
-
-type NoAuthOptions struct {
-	TenantID string
-}
-
-func (n *NoAuthOptions) GetTenantId() string {
-	return n.TenantID
 }
 
 // Config is a struct that defines some options for calling the Client.
