+++ osdsctl/cli/cli.go
@@ -29,21 +29,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const (
-	// Opensds Auth EVNs
-	OpensdsEndpoint     = "OPENSDS_ENDPOINT"
-	OpensdsAuthStrategy = "OPENSDS_AUTH_STRATEGY"
-	OpensdsTenantId     = "OPENSDS_TENANT_ID"
-
-	// Keystone Auth ENVs
-	OsAuthUrl      = "OS_AUTH_URL"
-	OsUsername     = "OS_USERNAME"
-	OsPassword     = "OS_PASSWORD"
-	OsTenantName   = "OS_TENANT_NAME"
-	OsProjectName  = "OS_PROJECT_NAME"
-	OsUserDomainId = "OS_USER_DOMAIN_ID"
-)
-
 var (
 	client      *c.Client
 	rootCommand = &cobra.Command{
@@ -69,28 +54,6 @@ func init() {
 	flags.BoolVar(&Debug, "debug", false, "shows debugging output.")
 }
 
-func LoadKeystoneAuthOptionsFromEnv() *c.KeystoneAuthOptions {
-	opt := c.NewKeystoneAuthOptions()
-	opt.IdentityEndpoint = os.Getenv(OsAuthUrl)
-	opt.Username = os.Getenv(OsUsername)
-	opt.Password = os.Getenv(OsPassword)
-	opt.TenantName = os.Getenv(OsTenantName)
-	projectName := os.Getenv(OsProjectName)
-	opt.DomainID = os.Getenv(OsUserDomainId)
-	if opt.TenantName == "" {
-		opt.TenantName = projectName
-	}
-	return opt
-}
-
-func LoadNoAuthOptionsFromEnv() *c.NoAuthOptions {
-	tenantId, ok := os.LookupEnv(OpensdsTenantId)
-	if !ok {
-		return c.NewNoauthOptions(constants.DefaultTenantId)
-	}
-	return c.NewNoauthOptions(tenantId)
-}
-
 type Writer struct{}
 
 // do nothing
@@ -104,14 +67,14 @@ func Run() error {
 		log.SetOutput(Writer{})
 	}
 
-	ep, ok := os.LookupEnv(OpensdsEndpoint)
+	ep, ok := os.LookupEnv(c.OpensdsEndpoint)
 	if !ok {
 		return fmt.Errorf("ERROR: You must provide the endpoint by setting " +
 			"the environment variable OPENSDS_ENDPOINT")
 	}
 	cfg := &c.Config{Endpoint: ep}
 
-	authStrategy, ok := os.LookupEnv(OpensdsAuthStrategy)
+	authStrategy, ok := os.LookupEnv(c.OpensdsAuthStrategy)
 	if !ok {
 		authStrategy = c.Noauth
 		fmt.Println("WARNING: Not found Env OPENSDS_AUTH_STRATEGY, use default(noauth)")
@@ -119,9 +82,9 @@ func Run() error {
 
 	switch authStrategy {
 	case c.Keystone:
-		cfg.AuthOptions = LoadKeystoneAuthOptionsFromEnv()
+		cfg.AuthOptions = c.LoadKeystoneAuthOptionsFromEnv()
 	case c.Noauth:
-		cfg.AuthOptions = LoadNoAuthOptionsFromEnv()
+		cfg.AuthOptions = c.LoadNoAuthOptionsFromEnv()
 	default:
 		cfg.AuthOptions = c.NewNoauthOptions(constants.DefaultTenantId)
 	}
